feat: introduce `Resolver` as wrapper around identity.rs
