+++ Cargo.toml
@@ -29,4 +29,5 @@ identity_stronghold = { git = "https://github.com/iotaledger/identity.rs.git", r
 ] }
 iota-sdk = { version = "1.0", features = ["stronghold"] }
 log = "0.4"
+signature = "2.2"
 tokio = { version = "1", features = ["full"] }
