+++ consumer/Cargo.toml
@@ -13,4 +13,5 @@ did_key = { path = "../did_key" }
 did_web = { path = "../did_web" }
 identity_iota.workspace = true
 iota-sdk.workspace = true
+signature.workspace = true
 tokio.workspace = true
