+++ src/lib.rs
@@ -1,2 +1,3 @@
+pub use consumer::resolver::Resolver;
 pub use producer::did_document::Method;
 pub use producer::SecretManager;
